+++ docs/design/method-body-inspection.md
@@ -339,8 +339,13 @@ read-only parallel cache consumption.
 `LibraryBodyAsyncSiblingSignatureMatcher` supplies the async-sibling
 subsystem's stateless signature decoding, exact identity/comparison, async
 return compatibility, optional cancellation matching, and bounded finding
-display. Reader-relative lookup, caches, accessibility, and dispatch
-suppression remain assembly-builder policy.
+display. `LibraryBodyAsyncSiblingDispatchAnalyzer` owns reader-relative type
+relationships, constructed generic projection, virtual-slot and MethodImpl
+correspondence, constrained-method suppression, and conservative unknown
+handling. It consumes assembly-builder callbacks for synchronized external
+type resolution and the shared per-type method-name index. Reader-relative
+candidate lookup, caches, accessibility, orchestration, diagnostics, and
+result ordering remain assembly-builder policy.
 `CallerUnsafeMode_PointerSignatureIsImplicitWhenModuleNotOptedIn`,
 `OptimizationOpportunities_AsyncStateMachine_IsAmortized`, and
 `Allocations_ClassifiesCrossAndInAssemblyValueTypeNewobj_ByShape` gate
